Revert "Use a partial wakelock, instead of keeping screen awake"
This reverts commit 29652268430148f2342cf976f28777c2bf1b0f4e.

Rather Partial wakelock doesnt stop onStop from triggering, so one
needs to move the required logic into a service.

For now avoiding that, as I may need to refactor some of the other
logics to also work through/along-with service.

This is more a simple experiemental code to cross check, so to hell
with device power for now.