Main: Shallow copy the eventFLog of SensorMa
So if the main array is manipulated wrt elements being added or
removed, I assume this shouldnt get affected, I need to check
the clone implementation wrt ArrayList later.

I am assuming the shallow copy is not too simple that only the
top level Arraylist meta data is duplicated, but hopefully the
list of pointers wrt each element in the Arraylist is also
duplicated (and inturn not beyond).