Main+:PlotData work with a duplicate copy which is replaced always
Maintain a duplicate of the eventFLog, which is created+updated by
holding the eventFLog related mutex.

Each time it is requested to be updated, a new list is created.

PlotData inturn works on this duplicated list of sensor events.
This should ensure that the copy, if being worked on by PlotData
doesnt get emptied out behind its back. HOWEVER the flow could
still lead to a situation where plotdata picks up the list as
it is being created+updated, so new elements may get added to
this list. Need to check, if this can create a problem with parallel
use of such a list (ie by update logic and plotdata logic).