Main:PlotData: AutoScale XAxis based on number of data records
But for some reason, it is not showing the data plot for the initial
few 10s of data records. So may revert for now.

Idea was to see the less number of data records, at the begining,
in a zoomed in format wrt x-axis, but there is some issue wrt Canvas
withTransform scaling.