Main+:PlotData not visible, why PlotData called too often?
As logic was updated to plot the data, only after a new eventslist
is got, the plotted data is no longer showing up, bcas the plotdata
composable is getting called more times, and the data is getting
plotted in only some of those calls.

Trying to see why this too often calls, Extending gap wrt refreshme
timer as well as Avoiding the refreshMe timer also doesnt seem to
have helped.